moved the Timer code into its own view
