+++ KnowYourAngles/PlayMode/PlayView.swift
@@ -17,20 +17,10 @@ struct PlayView: View {
     
     @StateObject var summary = SummaryGrid()
     @State var showSummary = false
-    
-    @State private var isEditing = false
-    @State private var minRemaining = 0
-    @State private var secRemaining = 0
-    
-    @State var timer = Timer.publish(every: 1, on: .main, in: .common).autoconnect()
-    @State var timerOn = false
-    @State private var wakeUp = Date()
-    @State var timerLabel = "Timer"
-    
+   
     @ViewBuilder
     var body: some View {
-        /*let timeFormatter = DateIntervalFormatter()
-         timeFormatter.timeStyle = .*/
+        
         ZStack {
             
             BackgroundColor()
@@ -39,81 +29,9 @@ struct PlayView: View {
                 
                 VStack (alignment: .center, spacing: 30, content: {
                     
-                    if(!modelData.finished)
-                    {
-                        HStack(alignment: .center) {
-                            Menu (timerLabel) {
-                                Button(action: {timerLabel = "00:30"}) {
-                                    Text("00:30")
-                                }
-                                
-                                Button(action: {timerLabel = "01:00"}) {
-                                    Text("01:00")
-                                }
-                                
-                                Button(action: {timerLabel = "01:30"}) {
-                                    Text("01:30")
-                                }
-                                
-                                Button(action: {timerLabel = "02:00"}) {
-                                    Text("02:00")
-                                }
-                                
-                                Button(action: {timerLabel = "02:30"}) {
-                                    Text("02:30")
-                                }
-                                
-                                Button(action: {timerLabel = "03:00"}) {
-                                    Text("03:00")
-                                }
-                                
-                                Button(action: {timerLabel = "03:30"}) {
-                                    Text("03:30")
-                                }
-                                
-                                Button(action: {timerLabel = "04:00"}) {
-                                    Text("04:00")
-                                }
-                                
-                                Button(action: {timerLabel = "04:30"}) {
-                                    Text("04:30")
-                                }
-                                
-                                Button(action: {timerLabel = "05:00"}) {
-                                    Text("05:00")
-                                }
-                            }
-                            .foregroundColor(Color.black)
-                            
-                            //Spacer()
-                            
-                            Image(systemName: timerOn ? "hourglass.tophalf.fill" : "hourglass.bottomhalf.fill")
-                                .foregroundColor(Color(red: 40.0/255, green: 204.0/255, blue: 198.0/255, opacity: 1.0))
-                                .onTapGesture {
-                                    if(timerLabel != "Timer")
-                                    {
-                                        timerOn.toggle()
-                                        if(timerOn)
-                                        {
-                                            minRemaining = Int(timerLabel.components(separatedBy: ":")[0]) ?? 0
-                                            secRemaining = Int(timerLabel.components(separatedBy: ":")[1]) ?? 0
-                                            startTimer()
-                                        }
-                                        else
-                                        {
-                                            stopTimer()
-                                        }
-                                    }
-                                }
-                        }
-                        .font(.system(size: 80))
-                        .padding(.leading, 0.0)
-                    }
-                    else
-                    {
-                        Text("00:00")
-                            .font(.system(size: 80))
-                    }
+                    TimerView()
+                        .scaledToFit()
+                    
                     ZStack {
                         ProgressSymbol(name: "KYA_Star_Icon")
                             .frame(width: 500, height: 500)
@@ -199,49 +117,12 @@ struct PlayView: View {
                     .font(.system(size: 36, weight: .regular, design: .serif))
                     .foregroundColor(Color(red: 127.0/255, green: 255.0/255, blue: 250.0/255, opacity: 1.0))
                 }
-                
                 Spacer()
             })//end VStack
             .frame(width: 500, alignment: .center)
             .padding()
         }//end Zstack
-        .onReceive(timer) { time in
-            //decrease the timer by one second
-            secRemaining -= 1
-            //if we've reached the end of a minute, subtract a minute and reset seconds to 59.
-            if(secRemaining == -1 && minRemaining > 0)
-            {
-                minRemaining -= 1;
-                secRemaining = 59;
-            }
-            if(secRemaining < 10)
-            {
-                timerLabel = "0" + String(minRemaining) + ":0" + String(secRemaining)
-            }
-            else
-            {
-                timerLabel = "0" + String(minRemaining) + ":" + String(secRemaining)
-            }
-            if(secRemaining <= 0 && minRemaining <= 0)
-            {
-                stopTimer()
-                timerOn = false
-                modelData.finished = true
-            }
-        }
-        .onAppear(perform: {
-            stopTimer()
-        })
     }//end View
-    
-    func stopTimer() {
-        self.timer.upstream.connect().cancel()
-    }
-    
-    func startTimer() {
-        timer = Timer.publish(every: 1, on: .main, in: .common).autoconnect()
-    }
-    
 }//end PlayView struct
 
 struct PlayView_Previews: PreviewProvider {
